refactor: dropdown 쿼리 파라미터를 쓰는 훅과 안쓰는 훅 분리 #448
